+++ jaorm-dsl/src/main/java/io/github/ulisse1996/jaorm/dsl/query/impl/SubQueryImpl.java
@@ -1,5 +1,6 @@
 package io.github.ulisse1996.jaorm.dsl.query.impl;
 
+import io.github.ulisse1996.jaorm.dsl.config.QueryConfig;
 import io.github.ulisse1996.jaorm.dsl.query.common.trait.WithSubQuerySupport;
 import io.github.ulisse1996.jaorm.entity.SqlColumn;
 
@@ -8,8 +9,8 @@ public class SubQueryImpl<T, R> extends SelectedImpl<T, R> implements WithSubQue
     private final String colum;
 
     @SuppressWarnings("unchecked")
-    public SubQueryImpl(SqlColumn<?, R> column) {
-        super((Class<T>) column.getEntity(), false);
+    public SubQueryImpl(SqlColumn<?, R> column, QueryConfig config) {
+        super((Class<T>) column.getEntity(), config);
         this.colum = column.getName();
     }