+++ jaorm-test/jaorm-test-core/src/main/java/io/github/ulisse1996/jaorm/integration/test/spring/SpringConfiguration.java
@@ -1,7 +1,0 @@
-package io.github.ulisse1996.jaorm.integration.test.spring;
-
-import org.springframework.boot.SpringBootConfiguration;
-
-@SpringBootConfiguration
-public class SpringConfiguration {
-}