+++ jaorm-test/jaorm-test-spring/src/test/java/io/github/ulisse1996/jaorm/integration/test/spring/repository/SpringApplicationDAO.java
@@ -1,7 +1,8 @@
-package io.github.ulisse1996.jaorm.integration.test.spring;
+package io.github.ulisse1996.jaorm.integration.test.spring.repository;
 
 import io.github.ulisse1996.jaorm.BaseDao;
 import io.github.ulisse1996.jaorm.annotation.Dao;
+import io.github.ulisse1996.jaorm.integration.test.spring.entity.SpringEntity;
 import org.springframework.stereotype.Repository;
 import org.springframework.web.context.annotation.ApplicationScope;