+++ CHANGELOG.md
@@ -9,6 +9,8 @@
 - Feature: Merge, Count and GeneratedKeys vendor specifics
 - Rework: IT Rework using Test Containers
 - Bugfix: Missing node name on Graph read
+- Bugfix: Missing upsert on single Entity update
+- Bugfix: Missing QueryConfig on SubQuery creation
 
 ## 1.6.0
 - Refactor: Move generation bundle SPI to SPI Provider for Single Feature